requirements.txt
FastAPI>=0.100.0
uvicorn>=0.23.0
SQLAlchemy>=2.0.0
alembic>=1.10.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
email-validator>=2.0.0
